requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.7.1
Django==5.2.1
django-appconf==1.1.0
django-autoslug==1.9.9
django-celery-email==3.0.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-environ==0.12.0
django-filter==25.1
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.16.0
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.1.0
flake8==7.1.2
flower==2.0.1
frozenlist==1.6.0
gunicorn==23.0.0
humanize==4.12.3
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
kombu==5.5.3
mccabe==0.7.0
multidict==6.4.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.2
platformdirs==4.3.8
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
tornado==6.4.2
twilio==9.6.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
